+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.1.0 (22/10/2020)
+
+## Changed
+
+- Added new default repository method Find Many `findMany(array $ids)`
+- Altered HasRepository method to contain a default location for repositories
+
 ## v1.0.1 (06/10/2020)
 
 ## Changed